requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
motor==3.3.2
PyJWT==2.8.0
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic-settings==2.1.0
bcrypt==4.1.2
fastapi-limiter==0.1.6
pytest==7.4.3
python-jose[cryptography]==3.3.0
pytest-asyncio==0.23.8